Play now an instance variable of State.  Related changes in Bowling, Frame, States.
One concern with the previous design was that Play was being passed
from Bowling to Frame to State, where it was finally used.  So I made
Play an instance variable of State, and passed it during creation of
State object.  This eliminated the long passage of Play through many
objects.